requirements.txt
pymongo
python-dotenv
pandas
fastapi
uvicorn
certifi